[Seat] : 코드 정리 #146
SeatCommonHelper
- seatPrefix
- makeCacheKey 메서드 추가
- createCacheBlockKey 메서드 추가

CreateCacheSeatsService
- seatPrefix 분리
- makeCacheKey 분리

GetCacheBlockSeatsService
- createCacheBlockKey 분리

ProcessBlockSeatsService
- createCacheBlockKey 분리

SeatFailedService
- SeatCommonHelper 적용

SeatService
- SeatCommonHelper 적용

SeatSucceedService
- SeatCommonHelper 적용

SeatListenerHelper
- createCacheBlockKey 분리
- makeCacheKey 분리
