+++ seat-service/src/main/java/com/boeingmerryho/business/seatservice/infrastructure/adaptor/kafka/listener/SeatFailedListener.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.seatservice.infrastructure.adaptor.response;
+package com.boeingmerryho.business.seatservice.infrastructure.adaptor.kafka.listener;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
