+++ seat-service/src/main/java/com/boeingmerryho/business/seatservice/infrastructure/helper/SeatCommonHelper.java
@@ -1,5 +1,7 @@
 package com.boeingmerryho.business.seatservice.infrastructure.helper;
 
+import java.time.LocalDate;
+
 import org.springframework.stereotype.Component;
 
 import com.boeingmerryho.business.seatservice.domain.Seat;
@@ -16,6 +18,8 @@ import lombok.extern.slf4j.Slf4j;
 public class SeatCommonHelper {
 	private final SeatRepository seatRepository;
 
+	public final String seatPrefix = "seat:";
+
 	public Seat getSeatById(Long id) {
 		return seatRepository.findById(id)
 			.orElseThrow(() -> new GlobalException(SeatErrorCode.NOT_FOUND_SEAT));
@@ -23,5 +27,29 @@ public class SeatCommonHelper {
 
 	public void save(Seat seat) {
 		seatRepository.save(seat);
+	}
+
+	public String makeCacheKey(Seat seat, LocalDate date) {
+		StringBuilder builder = new StringBuilder()
+			.append(seatPrefix)
+			.append(date)
+			.append(":")
+			.append(seat.getSeatBlock())
+			.append(":")
+			.append(seat.getSeatColumn())
+			.append(":")
+			.append(seat.getSeatRow());
+
+		return builder.toString();
+	}
+
+	public String createCacheBlockKey(Integer blockId, LocalDate date) {
+		StringBuilder cacheBlockKey = new StringBuilder()
+			.append(seatPrefix)
+			.append(date)
+			.append(":")
+			.append(blockId);
+
+		return cacheBlockKey.toString();
 	}
 }
